Made some changes for better LaTeX support. Added RefTeX changes
Also made a change for Emacs-Lisp because I like
the (un)comment-region function from AucTeX, so I brought it to that
mode, because it's likely that I'll need to (un)comment large regions
at a time.

Also ignoring a projectile file for some reason, I don't know why it's there.