Rust: Went through and figured out some extra packages to add
I actually skipped out on rust-mode, since it is not officially
supported by GNU Emacs, and the Rust team doesn't keep it up-to-date.
But, rustic is a fork of rust-mode that seems to behave much better.

I have set up rustic to also work with the LSP, and plug into the
rust-analyzer Cargo crate.

I am also toying around with racer for code completion and source code
navigation, though that one is a bit optional.