Personal: Make rainbow-mode pakcage available for any later configs
If I want to use rainbow-mode in a buffer later, I need to have it
pulled down with use-package somewhere before I use it.

I do that action in my personal-settings config file, which means it
is also a personal choice.
I have matched it with the other parenthetical changes I made, because
it works on parentheses, brackets and braces