Early Init: Started an early-init.el file
This file is evaluated BEFORE ANY (package-initialize) is evaluated.
In this file, I configure things that should be true before Emacs, the
editor, and package user is started.

So, I use early-init.el to increase the garbage collection threshold
of the interpreter, which is what starts eating things up and slowing
things down during boot.