Server: Updated so it doesn't try to run a new server every time
Before, every time my Emacs config was evaluated, it would attempt to
start a new server with the default name. This was problematic for the
use of Emacs with the client-server model with the Emacs daemon
running from user login. Now though, it won't start a new daemon
unless one is not already present.