Explicitly use straight to load project.el, very early in config time
This MUST happen here, because project.el is shipped with Emacs and
can be pulled in by other packages. If a package attempts to use
project.el, `(require 'project)` before I load it with straight, the
old verion shipped with Emacs will be used. Then, these newer packages
will attempt to use the old project.el, which causes no end of
problems.

The culprit here was the xref package.