Correct initial Emacs server frame to use correct font and size
If you manually set the face-attribute and the frame-font, the
intended changes do not take effect because the server does not
actually create any frames. Thus, the first frame that is created that
attaches to the server will have the default font and size, which is
either far too small or far too large.

Setting the `default-frame-alist` will mean my preferred font is the
default one to use on a new frame. The `initial-frame-alist` helps
Emacs understand where these fonts are (I think).

I am not sure baout the `initial-frame-alist` one, because I was
getting errors that Emacs could not find the font that I wanted to use
because it was evaluating the variable.