Use quaisquote to construct the frame-alists
Because quote prevents evaluation of the entire sexp, it means the
value actually in the alist was the symbol for my constant. But, we
need the evaulated version in the list. The frame-alists do *not*
evaluate their terms after being set. So we force evaluation with the
quasiquote, evaluating just the constant symbol and not the font
symbol.