Change how org-roam is loaded based on the system I am running on
The way this is done is SUPER hacky. I hate myself for having to do
this. But the problem comes down to the :straight symbol being used as
a plist and its quoting by straight.el. I can provide a function to
determine the backend to use, but the function MUST be one word (no
hyphens at all). There are likely other limitations that I am not
aware of.