Remove unneeded docker-config
If I am managing docker containers, I will probably do it from the
command-line rather than through Emacs. While it would be nice to do
this through Emacs, there is no reason to shoot myself in the foot
with poor tools for Docker, especially when its CLI is pretty decent.