Add ediff configuration
This configuration is based on Prot's ediff configuration.
See https://git.sr.ht/~protesilaos/dotfiles.

The most important changes are that splits are made to produce two
horizontally aligned windows (which contain buffers). The default is
to split the entire frame vertically, making 2 buffers stacked
vertically. This is confusing when ediffing source code.

The other major change is to put the ediff control buffer IN the same
frame as the files being ediffed. By default, the files being ediffed
are placed in the main frame and a tiny secondary frame is created for
the control buffer. This gets REALLY confusing. Instead, the control
buffer is kept in the same frame as the files being ediffed.

The other configurations are just to ease the use of ediff and prevent
myself from accidental edits while ediffing.