Require dired inside remove-file-whitespace function
Doing _does_ mean the first invocation will be slower, as the dired
library gets loaded into the Emacs process, but subsequent runs will
be just as fast as normal. On the plus side, this means that we do not
"pollute" the global namespace with dired functions if we do not need
them.