Use #' for local-set-key
local-set-key requires a function be passed as the second
argument (the function to call) and Emacs uses the #' symbol to quote
function names.

NOTE: Emacs allows the ' symbol too because under-the-hood, when the
keybinding is pressed, a (funcall 'the-func ...) is performed, and
funcall forcibly looks the symbol up in the function table. But this
is just my theory.