:straight built-in to :ensure nil for elpaca
elpaca does not have a notion of packages built-in to Emacs, since
the top-level elpaca form is just a macro around use-package. Setting
this to nil means that elpaca does not have to do any work, and that
the expanded version of the use-package will be used directly.

If you have enabled elpaca support for use-package, the following
happens:

(use-package pkg :ensure nil)

macroexpands to

((elpaca ... nil)
 (use-package pkg))

which just loads the package.