Use upstream racket-mode, rather than custom version
I wanted to have Guix and direnv manage my Racket installations on a
per-project basis. In particular, I wanted to be able to have multiple
versions on Racket on my system at one time. However, Racket is not
architected in a way that makes that easy. Instead, Racket uses a
single back-end binary which then provides multiple REPLs that
interact with it.

After discussion with racket-mode's author, we together came to the
conclusion that hooking envrc to set racket-mode's back-end would only
cause problems. As documented in the GitHub issue thread [1], and in
the documentation, the way to do this on a per-project basis is to
have it handled by .dir-locals.el using the following snippet:

  ((nil . ((eval . (racket-add-back-end default-directory)))))

[1] https://github.com/greghendershott/racket-mode/issues/706