Use emacsql-sqlite-builtin to make org-roam portable again
org-roam uses a sqlite database to cache cross-references between
notes/nodes. But to make the database fast enough for Emacs to use, we
had to compile a special sqlite for Emacs, which required a
C99-compliant compiler. Because I am on Guix and do not want to have
such a compiler present for everything, this caused issues.

The two easiest ways to solve this is by having Guix provide a version
of org-roam, or by having a manifest.scm which can compile
emacs-sqlite for you, and then leaving that environment once done
compiling. I chose to do the former option. But I am also updating the
manifest to include the compiler so we can come back and toy with
things later.

emacsql-sqlite-builtin uses features present in Emacs >=29 to avoid
the compilation step and instead just work with the database directly.
This makes things even MORE portable, because my Emacs config should
just work (TM). It also means I can stop special-casing my Guix
machines.