Use a hook to get the features I want
use-package runs :config after the feature symbol for that package is
available, NOT after the major-mode is loaded. In the macroexpanded
version, the eval-after-load form does not run after the major-mode is
active in the buffer. It runs after the feature is loaded, which is
distinct from when the major-mode (if the feature provides one) is
activated

Thus, all of my changes after the major-mode has done its thing,
namely in a :hook.