Force a full clone of hl-todo for magit-todos
magit-todos requires hl-todos version 1.9.0 or greater. However, the
maintainer of hl-todos does not want to include the "Version: X.Y.Z"
string in the main hl-todos.el file of the project. This is a fair
stance to take, because you would be duplicating version information
somewhere (Git history and the file).

However, elpaca RELIES on knowing the version number to make handle
dependency resolution. Elpaca can figure out the version on its own by
looking at the four fields, in this order:

  1. Via the package recipe's :version value, if any.
  2. In the package's -pkg.el file.
  3. In the package main elisp file's Version: header field
  4. The latest version git tag.

But because hl-todos does not include a version in one of the first 3
elements, it will try the fourth. But that will fail too, because
elpaca does a shallow, single-commit clone (:depth 1) of the project
to save space, which prevents branches and tags from being cloned.

By using :depth nil in hl-todos' use-package specification, we fetch
the full repository, which lets us get the Git tag version
information, and complete the dependency resolution.

[1] https://github.com/progfolio/elpaca/issues/229#issuecomment-1873490249
[2] https://github.com/alphapapa/magit-todos/issues/171#issuecomment-1934362142