Clean up Guix configuration stuff
guix-state-directory is a variable that I don't seem to need anymore?
I don't know, but having it set is not doing anything, so I just
removed it.

This guix package is the one that gives us major-modes for Guix
derivations and log files, so it is pretty important, even if I don't
use the interactive interface to control Guix stuff.