Remove org-mode's loading of languages for evaluation
I very rarely ever evaluate the source code in an Org-mode file and
use the document itself as a literate configuration or notebook.
Instead, I mostly use these for documentation and note-taking.

So let's remove the stuff for loading in a language's
interpreter/compiler for evaluating their source-code in the org-mode
buffer.