Grab TREE_SITTER_GRAMMAR_PATH to determine whether to use treesit
If I have control and access over a machine, then I will force myself
to use the env-var TREE_SITTER_GRAMMAR_PATH to determine where the
language's library actually is. I do this because Guix does this by
default and I want my Emacs config to be portable to other machines
too.

treesit-available-p returns non-nil on any machine that *technically*
supports tree-sitter, but is not a sufficient test to determine
whether to actually USE treesit.