Add TeX command to run texcount and open results in buffer
This will run the texcount binary on the main document, and produce
the results in an Emacs buffer. While the buffer will not be a compile
buffer (where a simple 'g' will re-run the texcount command without
further input, 'g' is bound to re-run the command, but it requires you
to confirm the command to use).