Configure TRAMP
There was a great blog post about how to make TRAMP feel faster [1].
This primarily boiled down to making the in-band vs. out-of-band
threshold higher and making sure that we use the direct async
processes when possible. This lets us create async processes on the
remote server, with far lower overhead than TRAMP originally had.

[1] https://coredumped.dev/2025/06/18/making-tramp-go-brrrr./