Remove cape-tex from list of Emacs capf's
The reason for this change is that sometimes \section in TeX/LaTeX
documents would ALWAYS be replaced by §. This was an inconsistent, but
reproducible issue. In particular, it depended on how fast I typed out
the "\sec" in \section (or if I made a typo). The completions that
were suggested were always § or \sectionmark. The more annoying part
was that if § was the suggested completion, then even if I MANUALLY
complete the \section, the § replacement ALWAYS happened.

The reason for this lies in a core design decision of corfu. In
particular, corfu is designed to only use built-in Emacs features, to
keep it leaner and integrate more tightly with Emacs itself. To gather
completion candidates (which are displayed in the little drop-down),
Corfu uses Emacs' built-in Completion At Point Functions (capf's).

Cape (Completion At Point Extensions) provide additional capf's for a
variety of things, including TeX/LaTeX. For TeX/LaTeX files/buffers,
cape-tex is a capf function that provides a completion that turns TeX
commands into a single unicode character.

I think the core problem is that there is a slightly racey problem
between capf's. Corfu runs through the list of functions in the
completion-at-point-function hook, stopping on the first one that
returns a completion. Listed below is the value of
completion-at-point-functions before this commit in a LaTeX buffer.

(TeX--completion-at-point
 cape-elisp-symbol
 cape-tex
 cape-keyword
 cape-history
 cape-elisp-block
 cape-file
 cape-dabbrev
 #f(compiled-function () #<bytecode 0xf0d9d5ba29d295>)
 LaTeX--arguments-completion-at-point
 ispell-completion-at-point)

I think TeX--completion-at-point is slow sometimes and doesn't provide
a replacement fast enough to "beat" cape-tex's unicode replacement.